streamlit_app/requirements.txt
streamlit
openai
supabase
python-docx
pypdf
tiktoken
python-dotenv
